move input toolbar to separate file
also need to move toolbarimport class to remove circular dependency
